Added minor files, included latest modifications
In initialize_real the modifications to the land use data, if urban
fraction values are available, have been commented out, as we wanted
to retain the original state. Originally, the land use was modified by
the urban fractions and the other way around.

In module_sf_urban, the plotting of sample values in the urban tiles was
changed to only look at categories >30. In the master code, only cat. 1,
which is urban in USGS, was regarded.